Enemy AI

+++ Source/ProjectO/Public/Items/Weapon.h
@@ -31,7 +31,7 @@ public:
 	void OnBoxOverlap(UPrimitiveComponent* OverlappedComponent, AActor* OtherActor, UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult);
 
 	UPROPERTY(VisibleAnywhere)
-	int32 Damage = 10;
+	float Damage = 10;
 
 	TArray<AActor*> IgnoreActors;
